GAPIT Tutorial
Blocks of code for different analysis methods.
Publications for each commented above the block.

For this study, I think the Compressed MLM (mixed linear model) will be the best. (Tutorial 1)

It's not as robust as FarmCPU but given our proof of concept approach and the data formatting issue, it works best.